+++ lib/antares.rb
@@ -7,6 +7,7 @@ module Antares
   class Error < StandardError; end
   class UnsupportedLexerError < Error; end
   class ResourceLimitError < Error; end
+  class GrammarError < Error; end
 
   def self.compatible?(lexer)
     klass = lexer.is_a?(Class) ? lexer : lexer.class
@@ -20,3 +21,8 @@ require_relative "antares/lexer_state_snapshot"
 require_relative "antares/lexer_driver"
 require_relative "antares/highlighter"
 require_relative "antares/structure"
+require_relative "antares/grammar"
+
+module Antares
+  private_constant :TMLanguageLexer, :TMLanguagePlist
+end